requirements.txt
a2wsgi==1.10.10
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.17.1
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.1
apache-airflow-core==3.1.1
apache-airflow-providers-common-compat==1.8.0
apache-airflow-providers-common-io==1.6.4
apache-airflow-providers-common-sql==1.28.2
apache-airflow-providers-google==18.1.0
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.1
apache-airflow-task-sdk==1.1.1
argcomplete==3.6.3
asgiref==3.10.0
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
cachetools==6.2.1
cadwyn==5.4.5
cattrs==25.3.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
db-dtypes==1.4.3
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.14
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gcloud-aio-auth==5.4.2
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.6.0
gcsfs==2025.10.0
google-ads==28.3.0
google-analytics-admin==0.26.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-cloud-aiplatform==1.124.0
google-cloud-alloydb==0.6.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-automl==2.17.0
google-cloud-batch==0.18.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-datatransfer==3.20.0
google-cloud-bigtable==2.34.0
google-cloud-build==3.33.0
google-cloud-compute==1.40.0
google-cloud-container==2.61.0
google-cloud-core==2.5.0
google-cloud-datacatalog==3.28.0
google-cloud-dataflow-client==0.10.0
google-cloud-dataform==0.7.0
google-cloud-dataplex==2.14.0
google-cloud-dataproc==5.23.0
google-cloud-dataproc-metastore==1.20.0
google-cloud-dlp==3.33.0
google-cloud-kms==3.7.0
google-cloud-language==2.18.0
google-cloud-logging==3.12.1
google-cloud-managedkafka==0.2.0
google-cloud-memcache==1.13.0
google-cloud-monitoring==2.28.0
google-cloud-orchestration-airflow==1.18.0
google-cloud-os-login==2.18.0
google-cloud-pubsub==2.33.0
google-cloud-redis==2.19.0
google-cloud-resource-manager==1.15.0
google-cloud-run==0.12.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.4.1
google-cloud-storage-transfer==1.18.0
google-cloud-tasks==2.20.0
google-cloud-texttospeech==2.33.0
google-cloud-translate==3.22.0
google-cloud-videointelligence==2.17.0
google-cloud-vision==3.11.0
google-cloud-workflows==1.19.0
google-crc32c==1.7.1
google-genai==1.47.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenback==1.2.1
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-gcp==0.2.2
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.5
linkify-it-py==2.0.3
litellm==1.76.3
lockfile==0.12.2
looker-sdk==25.18.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.7.0
natsort==8.4.0
numpy==2.3.4
oauthlib==3.3.1
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandas-gbq==0.29.2
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydata-google-auth==1.9.1
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pytest==8.4.2
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
PyYAML-ft==8.0.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
retryhttp==1.4.0
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.15.1
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.16.3
setproctitle==1.3.7
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlalchemy-bigquery==1.15.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.17.1
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.48.0
structlog==25.1.10
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
types-protobuf==6.32.1.20250918
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uritemplate==4.2.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wirerope==1.0.0
wrapt==2.0.0
yarl==1.22.0
zipp==3.23.0
